Rename bot classification to bot detection in the codebase to reflect the front-end content
