+++ src/_App/BotAnalysis/Detection/Bar.js
