+++ src/_App/BotAnalysis/Detection/Page.js
@@ -4,7 +4,7 @@ import React from 'react'
 import SectionIndex from "../SectionIndex"
 import Section from "./Section"
 
-export default function BotClassificationPage() {
+export default function BotDetectionPage() {
     return (
         <span>
             <SectionIndex/>
