+++ src/_App/sidebar.js
@@ -7,7 +7,7 @@ import CollectionTopics from './TweetCollection/Topics/Section'
 import CollectionResults from './TweetCollection/Results/Section'
 
 import BotAnalysis from './BotAnalysis/Page' // SectionIndex
-import BotClassification from './BotAnalysis/Classification/Section'
+import BotDetection from './BotAnalysis/Detection/Section'
 import BotNetworks from "./BotAnalysis/Networks/Section"
 //import BotActivity from './BotAnalysis/Activity/Section'
 import BotBeneficiaries from './BotAnalysis/Beneficiaries/Section'
@@ -30,7 +30,7 @@ var sidebar = [
             {"key": "bot-language",         "title": "Bot Language",        "component": BotLanguage},
             //{"key": "bot-activity",         "title": "Bot Activity",        "component": BotActivity},
             {"key": "bot-networks",         "title": "Bot Networks",        "component": BotNetworks},
-            {"key": "bot-detection",        "title": "Bot Detection",      "component": BotClassification}
+            {"key": "bot-detection",        "title": "Bot Detection",      "component": BotDetection}
         ]
     },
     {
